统一MQTT协议版本为枚举类型并优化内存管理
将协议版本从Byte类型统一为MqttVersion枚举，涉及核心类和测试用例，提升类型安全性和代码规范性。优化临时Byte数组的内存管理，采用Pool.Shared池化机制，减少GC压力。修复MQTT 5.0属性处理逻辑，确保判断和序列化一致。测试同步调整，保证兼容性和准确性。

+++ NewLife.MQTT/MqttClient.cs
@@ -453,7 +453,7 @@ public class MqttClient : DisposeBase
             Username = UserName,
             Password = Password,
             CleanSession = CleanSession,
-            ProtocolLevel = (Byte)Version,
+            ProtocolLevel = Version,
         };
 
         // 设置遗嘱消息
